Implement before_commit
